refactor(config): partition settings.cpp into domain partials
Extends src/config/settings/ with eight domain files and a settings_detail.h
holding the shared P_STORE macros and helpers. settings.cpp keeps the
load/save/reset core.
